fix: poc for using auth_config during push
Related to issue #189 this change uses the auth_config specifeid in the
build context when pushing a built docker_registry_image to the registry.
This is needed in the situationa where the registry is created by
terraform and one can therefore not preconfigure the provider registry
authentication. This change conflicts with the documentation update in
#314